requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http==4.9.0
astronomer-cosmos[dbt-bigquery]
protobuf
apache-airflow[postgres]
apache-airflow-providers-slack==3.0.0